Run prettier on backend and frontend
This cleans up the whitespace, line length, etc in a standard way.

+++ frontend/src/app/admin/design/components/filter/filter.component.ts
@@ -1,4 +1,3 @@
-
 import { Component, OnInit, Output, EventEmitter, Input } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { FormDialogComponent } from '../form-dialog/form-dialog.component';
@@ -6,57 +5,57 @@ import { FormDialogComponent } from '../form-dialog/form-dialog.component';
 @Component({
   selector: 'app-filter',
   templateUrl: './filter.component.html',
-  styleUrls: ['./filter.component.scss']
+  styleUrls: ['./filter.component.scss'],
 })
 export class FilterComponent implements OnInit {
-  @Output() edited: EventEmitter<any> = new EventEmitter()
-  @Output() onDelete: EventEmitter<boolean> = new EventEmitter()
-  dialogRef: MatDialogRef<any>
+  @Output() edited: EventEmitter<any> = new EventEmitter();
+  @Output() onDelete: EventEmitter<boolean> = new EventEmitter();
+  dialogRef: MatDialogRef<any>;
   form_data = [];
   icon: string;
   iconText: string;
-  pre: any
+  pre: any;
   baseform = [
     {
       name: 'component',
       label: 'Item Type',
       type: 'select',
       items: [
-        { name: "Label", value: "LabelComponent" },
-        { name: "Select", value: "SelectComponent" },
-        { name: "Search", value: "SearchComponent" },
-        { name: "Range", value: "RangeComponent" },
-        { name: "Date Range", value: "DateRangeComponent" },
+        { name: 'Label', value: 'LabelComponent' },
+        { name: 'Select', value: 'SelectComponent' },
+        { name: 'Search', value: 'SearchComponent' },
+        { name: 'Range', value: 'RangeComponent' },
+        { name: 'Date Range', value: 'DateRangeComponent' },
       ],
       required: true,
       onChange: (event) => {
         this.pre = event;
-        this.setFormDataOprions(event.value)
+        this.setFormDataOprions(event.value);
         this.dialogRef.close(true);
         this.openDialog();
       },
-    }
+    },
   ];
 
-
   setFormDataOprions(value) {
-
     switch (value) {
       case 'SearchComponent':
-        this.form_data = [...this.baseform, ...
-          [
+        this.form_data = [
+          ...this.baseform,
+          ...[
             {
               name: 'placeholder',
               label: 'Placeholder',
               type: 'text',
               required: true,
-            }
-          ]
-        ]
+            },
+          ],
+        ];
         break;
       case 'LabelComponent':
-        this.form_data = [...this.baseform, ...
-          [
+        this.form_data = [
+          ...this.baseform,
+          ...[
             {
               name: 'text',
               label: 'Text',
@@ -74,13 +73,14 @@ export class FilterComponent implements OnInit {
               label: 'Text',
               type: 'textarea',
               required: false,
-            }
-          ]
-        ]
+            },
+          ],
+        ];
         break;
       case 'RangeComponent':
-        this.form_data = [...this.baseform, ...
-          [
+        this.form_data = [
+          ...this.baseform,
+          ...[
             {
               name: 'source',
               label: 'Data Source',
@@ -92,13 +92,14 @@ export class FilterComponent implements OnInit {
               label: 'Placeholder',
               type: 'text',
               required: true,
-            }
-          ]
-        ]
+            },
+          ],
+        ];
         break;
       case 'DateRangeComponent':
-        this.form_data = [...this.baseform, ...
-          [
+        this.form_data = [
+          ...this.baseform,
+          ...[
             {
               name: 'source',
               label: 'Data Source',
@@ -110,13 +111,14 @@ export class FilterComponent implements OnInit {
               label: 'Placeholder',
               type: 'text',
               required: true,
-            }
-          ]
-        ]
+            },
+          ],
+        ];
         break;
       case 'SelectComponent':
-        this.form_data = [...this.baseform, ...
-          [
+        this.form_data = [
+          ...this.baseform,
+          ...[
             {
               name: 'source',
               label: 'Data Source',
@@ -128,9 +130,9 @@ export class FilterComponent implements OnInit {
               label: 'Placeholder',
               type: 'text',
               required: true,
-            }
-          ]
-        ]
+            },
+          ],
+        ];
         break;
       default:
         this.form_data = this.baseform;
@@ -141,47 +143,42 @@ export class FilterComponent implements OnInit {
   @Input() configs;
 
   delete() {
-    this.onDelete.emit(true)
+    this.onDelete.emit(true);
   }
 
   controls = [];
-  constructor(public dialog: MatDialog) { }
+  constructor(public dialog: MatDialog) {}
   setIcon() {
     let icons = {
       SelectComponent: 'list',
       LabelComponent: 'label',
       SearchComponent: 'search',
       RangeComponent: 'settings_ethernet',
-      DateRangeComponent: 'date_range'
-    }
+      DateRangeComponent: 'date_range',
+    };
     let iconsTexts = {
       SelectComponent: 'Select',
       LabelComponent: 'Label',
       SearchComponent: 'Search',
       RangeComponent: 'Range',
-      DateRangeComponent: 'Date Range'
-    }
-    this.icon = icons[this.configs.component]
+      DateRangeComponent: 'Date Range',
+    };
+    this.icon = icons[this.configs.component];
     this.iconText = iconsTexts[this.configs.component];
   }
   ngOnInit(): void {
     this.form_data = this.baseform;
     if (!this.configs.component) {
-
       this.openDialog();
-
     } else {
-      this.setFormDataOprions(this.configs.component)
-      this.setIcon()
+      this.setFormDataOprions(this.configs.component);
+      this.setIcon();
     }
-
-
-
   }
   editFilter: any = false;
   edit() {
-    this.editFilter = Object.assign({}, this.configs)
-    this.openDialog()
+    this.editFilter = Object.assign({}, this.configs);
+    this.openDialog();
   }
 
   openDialog(): void {
@@ -191,20 +188,20 @@ export class FilterComponent implements OnInit {
 
     this.dialogRef = this.dialog.open(FormDialogComponent, {
       width: '456px',
-      data: { form_data: this.form_data, configs: this.configs }
+      data: { form_data: this.form_data, configs: this.configs },
     });
 
-    this.dialogRef.afterClosed().subscribe(result => {
+    this.dialogRef.afterClosed().subscribe((result) => {
       if (result != true) {
-        if (result && result != true)
-          this.edited.emit(result)
-        if (!result && !this.editFilter)
-          this.delete()
+        if (result && result != true) this.edited.emit(result);
+        if (!result && !this.editFilter) this.delete();
         else if (this.editFilter != false && !result) {
-          this.edited.emit({ ...this.editFilter.componentConfigs, component: this.editFilter.component })
+          this.edited.emit({
+            ...this.editFilter.componentConfigs,
+            component: this.editFilter.component,
+          });
         }
       }
     });
   }
-
 }
